requirements.txt
streamlit
pyserial
sklearn
matplotlib
joblib
pandas
numpy
